Move router loaders and actions to feature hooks
- Extract boardsLoader to src/board/hooks/useBoardsLoader.ts
- Extract postDetailLoader to src/post/hooks/usePostDetailLoader.ts
- Extract createPostAction to src/post/hooks/useCreatePostAction.ts
- Update router imports to use feature-specific hooks
